[Release] Refresh OpenCode model cache before validating RELEASE_NOTES_MODEL
`opencode models` reads its catalog from ~/.cache/opencode/models.json and
falls back to the models.dev snapshot embedded in the binary when that file is
missing; the refresh it triggers is forked and never awaited. On a fresh CI
runner the first call therefore lists the snapshot of the pinned OpenCode
version, so a model released after that version looks unknown and the
release-notes job fails before doing any work.

Pass `--refresh` so the catalog is fetched up front (and the cache is warm for
the `opencode run` calls that follow), and reuse the check in the Slack message
script, which invoked `--model` with no validation at all.
